requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-crispy-forms==1.14.0
django-filter==22.1
django-jazzmin==2.6.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
yarg==0.1.9
